Fix migrations numbers for 32-bit systems
Migrations were indexed by timestamp. The problem is that PHP arrays are
indexed with integers. On a 32-bit system, the timestamp gets cast to an integer
when using it as an array index causing the indexes to be different from the
timestamps. We then loose are ability to migrate to a specific migration
via the file timestamp. By indexing with a string, this no longer matters